Move all application code into ./app
