+++ app/plugins/logging.js
