+++ app/routes/healthy.js
