+++ docker-compose.override.yaml
@@ -13,9 +13,8 @@ services:
       - "9229:9229"
     volumes:
       - ./.git/:/home/node/.git/
-      - ./server/:/home/node/server/
+      - ./app/:/home/node/app/
       - ./test/:/home/node/test/
       - ./test-output/:/home/node/test-output/
-      - ./index.js:/home/node/index.js
       - ./jest.config.js:/home/node/jest.config.js
       - ./package.json:/home/node/package.json
