+++ docker-compose.test.yaml
@@ -11,7 +11,7 @@ services:
     image: ffc-demo-api-gateway-development
     command: npm test
     volumes:
-      - ./test/:/home/node/app/test/
-      - ./test-output/:/home/node/app/test-output/
+      - ./test/:/home/node/test/
+      - ./test-output/:/home/node/test-output/
       - ./jest.config.js:/home/node/jest.config.js
       - ./package.json:/home/node/package.json
