+++ test/routes/healthy.test.js
@@ -3,7 +3,7 @@ describe('Healthy test', () => {
   let server
 
   beforeAll(async () => {
-    createServer = require('../../server')
+    createServer = require('../../app/server')
   })
 
   beforeEach(async () => {
