+++ test/routes/healthz.test.js
@@ -3,7 +3,7 @@ describe('Healthz test', () => {
   let server
 
   beforeAll(async () => {
-    createServer = require('../../server')
+    createServer = require('../../app/server')
   })
 
   beforeEach(async () => {
